tokio::sync::Semaphore is Sync if asked correctly
