+++ src/permit.rs
@@ -1,10 +1,3 @@
-use std::sync::{Mutex, TryLockError};
-use std::future::Future;
-use std::pin::Pin;
-use std::task::{Context, Poll};
-
-use crate::Canceled;
-
 use log::{warn, trace};
 
 #[cfg(feature = "tokio-semaphore")]
@@ -15,7 +8,8 @@ pub use tokio_semaphore::{
     blocking_permit_future,
     BlockingPermit,
     BlockingPermitFuture,
-    Semaphore
+    Semaphore,
+    SyncBlockingPermitFuture,
 };
 
 #[cfg(not(feature = "tokio-semaphore"))]
@@ -28,7 +22,8 @@ pub use intrusive::{
     blocking_permit_future,
     BlockingPermit,
     BlockingPermitFuture,
-    Semaphore
+    Semaphore,
+    SyncBlockingPermitFuture,
 };
 
 pub trait Semaphorish {
@@ -84,58 +79,12 @@ impl<'a> BlockingPermit<'a> {
     }
 }
 
-impl<'a> BlockingPermitFuture<'a> {
-    /// Make a `Sync` version of this future by wrapping with a `Mutex`.
-    pub fn make_sync(self) -> SyncBlockingPermitFuture<'a> {
-        SyncBlockingPermitFuture {
-            futr: Mutex::new(self)
-        }
-    }
-}
-
 impl<'a> Drop for BlockingPermit<'a> {
     fn drop(&mut self) {
         if self.entered.get() {
             trace!("Dropped BlockingPermit (semaphore)");
         } else {
             warn!("Dropped BlockingPermit (semaphore) was never entered")
-        }
-    }
-}
-
-/// A `Sync` wrapper available via [`BlockingPermitFuture::make_sync`].
-#[must_use = "must be `.await`ed or polled"]
-#[derive(Debug)]
-pub struct SyncBlockingPermitFuture<'a> {
-    futr: Mutex<BlockingPermitFuture<'a>>
-}
-
-impl<'a> SyncBlockingPermitFuture<'a> {
-    /// Construct given `Semaphore` reference.
-    pub fn new(semaphore: &'a Semaphore) -> SyncBlockingPermitFuture<'a>
-    {
-        SyncBlockingPermitFuture {
-            futr: Mutex::new(BlockingPermitFuture::new(semaphore))
-        }
-    }
-}
-
-impl<'a> Future for SyncBlockingPermitFuture<'a> {
-    type Output = Result<BlockingPermit<'a>, Canceled>;
-
-    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>)
-        -> Poll<Self::Output>
-    {
-        match self.futr.try_lock() {
-            Ok(mut guard) => {
-                let futr = unsafe { Pin::new_unchecked(&mut *guard) };
-                futr.poll(cx)
-            }
-            Err(TryLockError::Poisoned(_)) => Poll::Ready(Err(Canceled)),
-            Err(TryLockError::WouldBlock) => {
-                cx.waker().wake_by_ref(); //any spin should be brief
-                Poll::Pending
-            }
         }
     }
 }
